refactor: move operator runtime helpers
